+++ src/game2D/world/TileRenderer.java
@@ -38,8 +38,8 @@ public class TileRenderer {
 
 	public void render(World world, Camera camera) {
 
-		int posX = (int)camera.getPosition().x / (world.getScale() * 2);
-		int posY = (int)camera.getPosition().y / (world.getScale() * 2);
+		int posX = (int) camera.getPosition().x / (world.getScale() * 2);
+		int posY = (int) camera.getPosition().y / (world.getScale() * 2);
 		
 		for (int i = 0; i < world.getViewWidth(); i++) {
 			for (int j = 0; j < world.getViewHeight(); j++) {
@@ -61,9 +61,9 @@ public class TileRenderer {
 		if (tile_textures.containsKey(tile.getTexture())) {
 			tile_textures.get(tile.getTexture()).bind(0);
 		}
-		
-		int tile_x = 0; // tile.getX();
-		int tile_y = 0; // tile.getY();
+
+		float tile_x = 0.0f; // tile.getX();
+		float tile_y = 0.0f; // tile.getY();
 		
 		Matrix4f tile_pos = new Matrix4f().translate(new Vector3f(x * 2 + tile_x, y * 2 + tile_y, 0));
 		Matrix4f target = new Matrix4f();
